Fixed Circular Link Bugs

+++ src/cfd_mesh/CfdMeshMgr.cpp
@@ -4083,7 +4083,7 @@ void CfdMeshMgrSingleton::LoadDrawObjs( vector< DrawObj* > &draw_obj_vec )
 
     // Calculate constants for color sequence.
     const int ncgrp = 6; // Number of basic colors
-    const int ncstep = ceil((double)num_tags/(double)ncgrp);
+    const int ncstep = (int)ceil((double)num_tags/(double)ncgrp);
     const double nctodeg = 360.0/(ncgrp*ncstep);
 
     char str[256];
